requirements.txt
# Unified requirements for notebooks and Django deployment
# Pin major versions loosely to avoid CUDA/OS mismatches

# Core scientific stack
numpy
pandas
matplotlib
seaborn
scikit-learn
imbalanced-learn
tqdm

# PyTorch + vision (install will pick CPU or CUDA wheels appropriate for your system)
torch
torchvision

# Computer vision
opencv-python
Pillow

# Web app and REST API
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
